+++ packages/client/CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 
+- Deprecate TypeScript interface `NetworkNodeConfig`
+
 ### Removed
 
 ### Fixed
@@ -23,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 
-- Deprecate TypeScript interfaces `StrictStreamrClientConfig`
+- Deprecate TypeScript interface `StrictStreamrClientConfig`
 - Deprecate `gasPriceStrategy` config option in `contracts.ethereumNetworks`, use `highGasPriceStrategy` instead
 - Deprecate method parameter of `.waitForStorage()`